+++ TODO
@@ -103,6 +103,7 @@
     [ ] More detailed deployment instructions
 [x] Fix email on production
 [-] Create csv output of donations
+    [ ] Try to move to admin.py
     [ ] Put in admin template
     [ ] Add link in main admin page
 [ ] Add message to donation notification
